regorg specifyers
