+++ hydromt/models/_region/__init__.py
@@ -1,4 +1,4 @@
 """Model for dealing with region specifcation and manipulation."""
-from hydromt.region.region import Region
+from hydromt.models._region.region import Region
 
 __all__ = ["Region"]
