+++ hydromt/models/api.py
@@ -31,7 +31,7 @@ from hydromt.data_catalog import DataCatalog
 from hydromt.gis.raster import GEO_MAP_COORD
 from hydromt.io import configread
 from hydromt.io.writers import configwrite
-from hydromt.region import Region
+from hydromt.models import Region
 
 __all__ = ["Model"]
